pgbench: support multiple connection parameters
Treat repeated flags for -U, -h, -p as defining lists of the relevant objects.

This allows for workloads in which you can define multiple sets of users (say) to connect to the
database in question to simulate workloads with multiple types of access.

In this version of the feature, there is a hard-coded limit of 10 of each param type, but we could
presumably either lift this limitation or make the allocation of these params more dynamic.

As written, this is not particularly thread- or fork-safe, however the tables are initialized by the
time we fork or create threads, so the worst thing that happens is that some of the connection lists
might not be used.  Since the point of this feature at this stage is just to vary the connected
users this is probably sufficient.

Quick implementation note: the new `param_set` struct serves to implement a round-robin interface
for just returning the "next" item in the list of items that have been added to it.  It is safe to
use with NULL items, or with no items at all, in which case the GetNextParamSet() routine will just
return a NULL pointer.

Since we support multiple params with this `param_set`, each with its own independent counters, if
there are different numbers of items in each, each call to doConnect() will use the next value for
each of them; i.e., if called frequently enough you will end up with different permutations as each
item round-robins around.