Support CREATE ROLE IF NOT EXISTS for roles, users, and groups
By choosing to implement CREATE ROLE IF NOT EXISTS instead of CREATE OR REPLACE ROLE, we can return
an existing oid while not throwing an error when using this form of the CREATE ROLE statement.  This
does mean that if writing code which expects a specific state for the ROLE with specific attributes,
you will likely want to separate these out into two separate statements:

    CREATE ROLE IF NOT EXISTS my_role;
    ALTER ROLE my_role WITH INHERIT SUPERUSER CREATEDB;

Instead of:

    CREATE ROLE IF NOT EXISTS my_role WITH INHERIT SUPERUSER CREATEDB;

While the second form will DWYM in cases where the role does not exist, it will not change the
existing object to reflect the role permissions passed in, and will instead keep the existing
permissions.  This is a conscious choice at this point.

What this *does* mean is that you can avoid issues caused by needing to (e.g.) DROP and immediately
CREATE a user/role when the intent is just "make sure this role exists".