Add cluster-wide Page Features
Page features are a standardized way of assigning and using dynamic space usage from the tail end of
a disk page.  These features are set at cluster init time (so configured via `initdb` and
initialized via the bootstrap process) and affect all disk pages.

A PageFeatureSet is effectively a bitflag of all configured features, each of which has a fixed
size.  If not using any PageFeatures, the storage overhead of this is 0.

Rather than using a variable location struct, an implementation of a PageFeature is responsible for
an offset and a length in the page.  The current API returns only a pointer to the page location for
the implementation to manage, and no further checks are done to ensure that only the expected memory
is accessed.

Access to the underlying memory is synonymous with determining whether a given cluster is using an
underlying PageFeature, so code paths can do something like:

    char *loc;

    if ((loc = ClusterGetPageFeatureOffset(page, PF_MY_FEATURE_ID)))
    {
        // ipso facto this feature is enabled in this cluster *and* we know the memory address
        ...
    }

Since this is direct memory access to the underlying Page, ensure the buffer is pinned.  Explicitly
locking (assuming you stay in your lane) should only need to guard against access from other
backends of this type if using shared buffers, so will be use-case dependent.

This does have a runtime overhead due to moving some offset calculations from compile time to
runtime. It is thought that the utility of this feature will outweigh the costs here.

Candidates for page features include 32-bit or 64-bit checksums, encryption tags, or additional
per-page metadata.

While we are not currently getting rid of the pd_checksum field, this mechanism could be used to
free up that 16 bits for some other purpose. One such purpose might be to mirror the cluster-wise
PageFeatureSet, currently also a uint16, which would mean the entirety of this scheme could be
reflected in a given page, opening up per-relation or even per-page setting/metadata here.  (We'd
presumably need to snag a pd_flags bit to interpret pd_checksum that way, but it would be an
interesting use.)