fix: use the correct permanence for pg_class to when scanning
The IVs being constructed here were using the wrong value for pg_class, which itself is a permanent
relation, which meant that creating a new database on this was forcing the DecryptPage to use the
wrong vectors, which caused page failures when creating a page using this method.

If the page was already in shared buffers, this would not trigger this issue, so connecting to
template0 then disconnecting and running a new create statement would not result in the same error.