Add page feature for 64-bit checksums
Since we reclaimed the space from the pd_checksums field for storing a page's
features, we present the use of a 64-bit page checksum as an alternative.  This
uses an arbitrarily chosen 64-bit hash for demo purposes (TBD: is this
compatible, or do we need a replacement?) to demonstrate the use of this feature.

Since one of the main motivators of page features is to provide space for
authenticated page encryption, we make this optional in order to ensure that
either the 64-bit checksum (this patch) or the 64-bit authtag (future patch)
will live in the final 8 bytes of the page at a single constant offset,
hopefully allowing other programs that need to know how to handle the new format
to do so in a much easier way.