Make the cluster cipher pluggable in code, add XTS
Rather than use an explicit cipher, we add a symbol for PG_CIPHER_DEFAULT and
the various infra required for supporting different cipher types.  As part of
this, we illustrate by adding XTS, which also requires us supporting a separate
for symbol for the MGR_KEK_KEY_LEN due to differing KEK and DEK lengths when
using XTS support.