Add reserved_page_space to Page structure
This space is reserved for extended data on the Page structure which will be
ultimately used for encrypted data, extended checksums, and potentially other
things.  This data appears at the end of the Page, after any `pd_special` area,
and will be calculated at runtime based on specific ControlFile features.

No effort is made to ensure this is backwards-compatible with existing clusters
for `pg_upgrade`, as we will require logical replication to move data into a
cluster with different settings here.

In order to have decent performance, we expand the approach used with the
variable blocksize patch and continue using nested inline switch statements with
calculated constants, adding a new enum/parameter for the allowed/reserved page
sizes.  We support a specific number of buckets, so the reserved space must be a
known multiple of the bucket size (8 bytes per bucket in this patch, with up to
2 buckets possible).  Adjusting the bucket size is a trivial recompilation, and
adding support for additional buckets is a straightforward copy/paste
adjustment. (I fiddled for a bit to get the preprocessor to handle the
repetition here in the inline static function, but wasn't quite able to crack
it.)

We store the total reserved space in the ControlFile rather than the enum used
since the actual enum-based bucket implementation is an implementation detail;
as long as we can properly handle the number of bytes being reserved then we're
golden.  This gives us some flexibility on implementation going forward as long
as we can respect the same value here.