fix: add a 3rd bucket for reserved space
Based on testing, with -O2 this is the point at which all existing inline
functions continue to be inlined; more than this and they start to be created as
independent functions (though not all instances for reasons not currently clear
to me).  Note that the number of buckets is what is important here, and not the
actual size, which is independent of the actual values of the buckets.